Move Features class to Facades namespace
