+++ packages/framework/resources/views/components/post/article.blade.php
@@ -1,6 +1,6 @@
 <article aria-label="Article" id="{{ Hyde::url("posts/$page->identifier", '') }}" itemscope
          itemtype="http://schema.org/Article"
-        @class(['post-article mx-auto prose dark:prose-invert', 'torchlight-enabled' => Hyde\Helpers\Features::hasTorchlight()])>
+        @class(['post-article mx-auto prose dark:prose-invert', 'torchlight-enabled' => Hyde\Facades\Features::hasTorchlight()])>
     <meta itemprop="identifier" content="{{ $page->identifier }}">
     @if(Hyde::hasSiteUrl())
         <meta itemprop="url" content="{{ Hyde::url('posts/' . $page->identifier) }}">
