+++ packages/framework/src/Facades/Features.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Helpers;
+namespace Hyde\Facades;
 
 use Hyde\Framework\Services\DiscoveryService;
 use Hyde\Hyde;
@@ -10,6 +10,8 @@ use Hyde\Support\Concerns\JsonSerializesArrayable;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Str;
 use JsonSerializable;
+use function count;
+use function str_starts_with;
 
 /**
  * Allows features to be enabled and disabled in a simple object-oriented manner.
