+++ packages/framework/src/Foundation/PageCollection.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Foundation;
 
+use Hyde\Facades\Features;
 use Hyde\Foundation\Concerns\BaseFoundationCollection;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Exceptions\FileNotFoundException;
-use Hyde\Helpers\Features;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\HtmlPage;
