+++ packages/framework/src/Framework/Features/DataCollections/DataCollectionServiceProvider.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Features\DataCollections;
 
+use Hyde\Facades\Features;
 use Hyde\Framework\Features\DataCollections\Facades\MarkdownCollection;
-use Hyde\Helpers\Features;
 use Hyde\Hyde;
 use Illuminate\Foundation\AliasLoader;
 use Illuminate\Support\ServiceProvider;
